Initial idea for media upload provider
