+++ packages/story-editor/src/components/layout/index.js
@@ -44,6 +44,7 @@ import LayoutProvider from '../../app/layout/layoutProvider';
 import { ChecklistCheckpointProvider } from '../checklist';
 import { RightClickMenuProvider } from '../../app/rightClickMenu';
 import RightClickMenu from '../canvas/rightClickMenu';
+import { MediaUploadProvider } from '../../app/mediaUpload';
 
 const Editor = withOverlay(styled.section.attrs({
   'aria-label': __('Web Stories Editor', 'web-stories'),
@@ -85,20 +86,25 @@ function Layout({ header, children }) {
     <>
       <LayoutProvider>
         <ChecklistCheckpointProvider>
-          <HighlightsProvider>
-            <Editor zIndex={3}>
-              <CanvasProvider>
-                <RightClickMenuProvider>
-                  <Area area="lib">
-                    <Library />
-                  </Area>
-                  <Workspace header={header} />
-                  <RightClickMenu />
-                </RightClickMenuProvider>
-              </CanvasProvider>
-              {children}
-            </Editor>
-          </HighlightsProvider>
+          <MediaUploadProvider
+            title={__('Select media', 'web-stories')}
+            buttonInsertText={__('Insert media', 'web-stories')}
+          >
+            <HighlightsProvider>
+              <Editor zIndex={3}>
+                <CanvasProvider>
+                  <RightClickMenuProvider>
+                    <Area area="lib">
+                      <Library />
+                    </Area>
+                    <Workspace header={header} />
+                    <RightClickMenu />
+                  </RightClickMenuProvider>
+                </CanvasProvider>
+                {children}
+              </Editor>
+            </HighlightsProvider>
+          </MediaUploadProvider>
         </ChecklistCheckpointProvider>
       </LayoutProvider>
       <Snackbar.Container {...snackbarState} />
